Rename `components/core` to `components/services`
