+++ src/components/services/EdgeCoreManager.js
