+++ src/components/services/EdgeWalletCallbackManager.js
