+++ src/components/services/EdgeWalletsCallbackManager.js
